Remove zero-width unicode characters from incoming message
medic/cht-core#7654
